+++ lib/src/pick_int.dart
@@ -1,14 +1,17 @@
 import 'package:deep_pick/src/pick.dart';
 
-extension IntPick on RequiredPick {
+extension NullableIntPick on Pick {
+  @Deprecated('Use .asIntOrThrow()')
+  int Function() get asInt => asIntOrThrow;
+
   /// Returns the picked [value] as [int]
   ///
   /// {@template Pick.asInt}
   /// Parses the picked value as [int]. Also tries to parse [String] as [int]
   /// via [int.tryParse]
   /// {@endtemplate}
-  int asInt() {
-    final value = this.value;
+  int _parse() {
+    final value = required().value;
     if (value is int) {
       return value;
     }
@@ -18,14 +21,9 @@ extension IntPick on RequiredPick {
         return parsed;
       }
     }
-    throw PickException('value $value of type ${value.runtimeType} '
-        'at location ${location()} can not be parsed as int');
+    throw PickException(
+        'Type ${value.runtimeType} of $debugParsingExit can not be parsed as int');
   }
-}
-
-extension NullableIntPick on Pick {
-  @Deprecated('Use .asIntOrThrow()')
-  int Function() get asInt => asIntOrThrow;
 
   /// Returns the picked [value] as [int] or throws
   ///
@@ -33,7 +31,7 @@ extension NullableIntPick on Pick {
   int asIntOrThrow() {
     withContext(requiredPickErrorHintKey,
         'Use asIntOrNull() when the value may be null/absent at some point (int?).');
-    return required().asInt();
+    return _parse();
   }
 
   /// Returns the picked [value] as [int?] or returns `null` when the picked
@@ -43,7 +41,7 @@ extension NullableIntPick on Pick {
   int /*?*/ asIntOrNull() {
     if (value == null) return null;
     try {
-      return required().asInt();
+      return _parse();
     } catch (_) {
       return null;
     }