+++ test/src/pick_datetime_test.dart
@@ -23,14 +23,8 @@ void main() {
         expect(
           () => nullPick().asDateTimeOrThrow(),
           throwsA(pickException(containing: [
-            'required value at location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asDateTimeOrNull() when the value may be null/absent at some point (DateTime?).'
+            'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asDateTimeOrNull() when the value may be null/absent at some point (DateTime?).'
           ])),
-        );
-
-        expect(
-          () => nullPick().asDateTimeOrThrow(),
-          throwsA(
-              pickException(containing: ['unknownKey', 'null', 'DateTime'])),
         );
       });
 
@@ -38,13 +32,13 @@ void main() {
         expect(
           () => pick(Object()).asDateTimeOrThrow(),
           throwsA(pickException(containing: [
-            "value Instance of 'Object' of type Object at location \"<root>\" in pick(<root>) can not be parsed as DateTime"
+            'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as DateTime'
           ])),
         );
         expect(
           () => pick('Bubblegum').asDateTimeOrThrow(),
           throwsA(
-              pickException(containing: ['Bubblegum', 'String', 'DateTime'])),
+              pickException(containing: ['String', 'Bubblegum', 'DateTime'])),
         );
       });
     });
@@ -57,7 +51,7 @@ void main() {
         // ignore: deprecated_member_use_from_same_package
         () => nullPick().asDateTime(),
         throwsA(pickException(containing: [
-          'required value at location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asDateTimeOrNull() when the value may be null/absent at some point (DateTime?).'
+          'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asDateTimeOrNull() when the value may be null/absent at some point (DateTime?).'
         ])),
       );
     });
@@ -74,6 +68,71 @@ void main() {
 
       test('wrong type returns null', () {
         expect(pick(Object()).asDateTimeOrNull(), isNull);
+      });
+    });
+  });
+
+  group('pick().required().asDateTime*', () {
+    group('asDateTimeOrThrow', () {
+      test('parse String', () {
+        expect(pick('2012-02-27 13:27:00').required().asDateTimeOrThrow(),
+            DateTime(2012, 2, 27, 13, 27));
+        expect(
+            pick('2012-02-27 13:27:00.123456z').required().asDateTimeOrThrow(),
+            DateTime.utc(2012, 2, 27, 13, 27, 0, 123, 456));
+      });
+
+      test('parse DateTime', () {
+        final time = DateTime.utc(2012, 2, 27, 13, 27, 0, 123, 456);
+        expect(pick(time.toString()).required().asDateTimeOrThrow(), time);
+        expect(pick(time).required().asDateTimeOrThrow(), time);
+      });
+
+      test('null throws', () {
+        expect(
+          () => nullPick().required().asDateTimeOrThrow(),
+          throwsA(pickException(containing: [
+            'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent.'
+          ])),
+        );
+      });
+
+      test('wrong type throws', () {
+        expect(
+          () => pick(Object()).required().asDateTimeOrThrow(),
+          throwsA(pickException(containing: [
+            'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as DateTime'
+          ])),
+        );
+        expect(
+          () => pick('Bubblegum').required().asDateTimeOrThrow(),
+          throwsA(
+              pickException(containing: ['String', 'Bubblegum', 'DateTime'])),
+        );
+      });
+    });
+
+    test('deprecated asDateTime forwards to asDateTimeOrThrow', () {
+      // ignore: deprecated_member_use_from_same_package
+      expect(pick('2012-02-27 13:27:00').required().asDateTime(),
+          DateTime(2012, 2, 27, 13, 27));
+      expect(
+        // ignore: deprecated_member_use_from_same_package
+        () => nullPick().required().asDateTime(),
+        throwsA(pickException(containing: [
+          'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent.'
+        ])),
+      );
+    });
+
+    group('asDateTimeOrNull', () {
+      test('parse String', () {
+        expect(pick('2012-02-27 13:27:00').required().asDateTimeOrNull(),
+            DateTime(2012, 2, 27, 13, 27));
+      });
+
+      test('wrong type returns null', () {
+        expect(pick(Object()).required().asDateTimeOrNull(), isNull);
       });
     });
   });