+++ test/src/pick_int_test.dart
@@ -20,7 +20,7 @@ void main() {
         expect(
           () => nullPick().asIntOrThrow(),
           throwsA(pickException(containing: [
-            'required value at location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asIntOrNull() when the value may be null/absent at some point (int?).'
+            'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asIntOrNull() when the value may be null/absent at some point (int?).'
           ])),
         );
       });
@@ -29,13 +29,13 @@ void main() {
         expect(
           () => pick(Object()).asIntOrThrow(),
           throwsA(pickException(containing: [
-            "value Instance of 'Object' of type Object at location \"<root>\" in pick(<root>) can not be parsed as int"
+            'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as int'
           ])),
         );
 
         expect(
           () => pick('Bubblegum').asIntOrThrow(),
-          throwsA(pickException(containing: ['Bubblegum', 'String', 'int'])),
+          throwsA(pickException(containing: ['String', 'Bubblegum', 'int'])),
         );
       });
     });
@@ -47,7 +47,7 @@ void main() {
         // ignore: deprecated_member_use_from_same_package
         () => pick(Object()).asInt(),
         throwsA(pickException(containing: [
-          "value Instance of 'Object' of type Object at location \"<root>\" in pick(<root>) can not be parsed as int"
+          'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as int'
         ])),
       );
     });
@@ -63,6 +63,65 @@ void main() {
 
       test('wrong type returns null', () {
         expect(pick(Object()).asIntOrNull(), isNull);
+      });
+    });
+  });
+
+  group('pick().required().asInt*', () {
+    group('asIntOrThrow', () {
+      test('parse Int', () {
+        expect(pick(1).required().asIntOrThrow(), 1);
+        expect(pick(35).required().asIntOrThrow(), 35);
+      });
+
+      test('parse int String', () {
+        expect(pick('1').required().asIntOrThrow(), 1);
+        expect(pick('123').required().asIntOrThrow(), 123);
+      });
+
+      test('null throws', () {
+        expect(
+          () => nullPick().required().asIntOrThrow(),
+          throwsA(pickException(containing: [
+            'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent.'
+          ])),
+        );
+      });
+
+      test('wrong type throws', () {
+        expect(
+          () => pick(Object()).required().asIntOrThrow(),
+          throwsA(pickException(containing: [
+            'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as int'
+          ])),
+        );
+
+        expect(
+          () => pick('Bubblegum').required().asIntOrThrow(),
+          throwsA(pickException(containing: ['String', 'Bubblegum', 'int'])),
+        );
+      });
+    });
+
+    test('deprecated asInt forwards to asIntOrThrow', () {
+      // ignore: deprecated_member_use_from_same_package
+      expect(pick('1').asInt(), 1.0);
+      expect(
+        // ignore: deprecated_member_use_from_same_package
+        () => pick(Object()).asInt(),
+        throwsA(pickException(containing: [
+          'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be parsed as int'
+        ])),
+      );
+    });
+
+    group('asIntOrNull', () {
+      test('parse String', () {
+        expect(pick('2012').required().asIntOrNull(), 2012);
+      });
+
+      test('wrong type returns null', () {
+        expect(pick(Object()).required().asIntOrNull(), isNull);
       });
     });
   });