+++ test/src/pick_list_test.dart
@@ -7,14 +7,15 @@ void main() {
   group('pick().asList*', () {
     group('asListOrThrow', () {
       test('pipe through List', () {
-        expect(pick([1, 2, 3]).asListOrThrow((it) => it.asInt()), [1, 2, 3]);
+        expect(pick([1, 2, 3]).asListOrThrow((it) => it.asIntOrThrow()),
+            [1, 2, 3]);
       });
 
       test('null throws', () {
         expect(
-          () => nullPick().asListOrThrow((it) => it.asString()),
+          () => nullPick().asListOrThrow((it) => it.asStringOrThrow()),
           throwsA(pickException(containing: [
-            'required value at location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asListOrEmpty()/asListOrNull() when the value may be null/absent at some point (List<String>?).'
+            'Expected a non-null value but location "unknownKey" in pick(json, "unknownKey" (absent)) is absent. Use asListOrEmpty()/asListOrNull() when the value may be null/absent at some point (List<String>?).'
           ])),
         );
       });
@@ -79,21 +80,21 @@ void main() {
             {'asdf': 'Daenerys Targaryen'}, // <-- missing name key
           ]).asListOrThrow((pick) => Person.fromPick(pick)),
           throwsA(pickException(containing: [
-            'required value at location list index 1 in pick(json, 1 (absent), "name") is absent. Use asListOrEmpty()/asListOrNull() when the value may be null/absent at some point (List<Person>?).'
+            'Expected a non-null value but location list index 1 in pick(json, 1 (absent), "name") is absent. Use asListOrEmpty()/asListOrNull() when the value may be null/absent at some point (List<Person>?).'
           ])),
         );
       });
 
       test('wrong type throws', () {
         expect(
-          () => pick('Bubblegum').asListOrThrow((it) => it.asString()),
+          () => pick('Bubblegum').asListOrThrow((it) => it.asStringOrThrow()),
           throwsA(pickException(
-              containing: ['Bubblegum', 'String', 'List<dynamic>'])),
+              containing: ['String', 'Bubblegum', 'List<dynamic>'])),
         );
         expect(
-          () => pick(Object()).asListOrThrow((it) => it.asString()),
+          () => pick(Object()).asListOrThrow((it) => it.asStringOrThrow()),
           throwsA(pickException(containing: [
-            'value Instance of \'Object\' of type Object at location "<root>" in pick(<root>) can not be casted to List<dynamic>'
+            'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be casted to List<dynamic>'
           ])),
         );
       });
@@ -106,22 +107,23 @@ void main() {
         // ignore: deprecated_member_use_from_same_package
         () => pick(Object()).asList(),
         throwsA(pickException(containing: [
-          'value Instance of \'Object\' of type Object at location "<root>" in pick(<root>) can not be casted to List<dynamic>'
+          'Type Object of picked value "Instance of \'Object\'" using pick(<root>) can not be casted to List<dynamic>'
         ])),
       );
     });
 
     group('asListOrEmpty', () {
       test('pick value', () {
-        expect(pick([1, 2, 3]).asListOrEmpty((it) => it.asInt()), [1, 2, 3]);
+        expect(pick([1, 2, 3]).asListOrEmpty((it) => it.asIntOrThrow()),
+            [1, 2, 3]);
       });
 
       test('null returns null', () {
-        expect(nullPick().asListOrEmpty((it) => it.asInt()), []);
+        expect(nullPick().asListOrEmpty((it) => it.asIntOrThrow()), []);
       });
 
       test('wrong type returns empty', () {
-        expect(pick(Object()).asListOrEmpty((it) => it.asInt()), []);
+        expect(pick(Object()).asListOrEmpty((it) => it.asIntOrThrow()), []);
       });
 
       test('map empty list to empty list', () {
@@ -191,7 +193,7 @@ void main() {
             {'asdf': 'Daenerys Targaryen'}, // <-- missing name key
           ]).asListOrEmpty((pick) => Person.fromPick(pick)),
           throwsA(pickException(containing: [
-            'required value at location list index 1 in pick(json, 1 (absent), "name") is absent.'
+            'Expected a non-null value but location list index 1 in pick(json, 1 (absent), "name") is absent.'
           ])),
         );
       });
@@ -199,15 +201,16 @@ void main() {
 
     group('asListOrNull', () {
       test('pick value', () {
-        expect(pick([1, 2, 3]).asListOrNull((it) => it.asInt()), [1, 2, 3]);
+        expect(
+            pick([1, 2, 3]).asListOrNull((it) => it.asIntOrThrow()), [1, 2, 3]);
       });
 
       test('null returns null', () {
-        expect(nullPick().asListOrNull((it) => it.asInt()), isNull);
+        expect(nullPick().asListOrNull((it) => it.asIntOrThrow()), isNull);
       });
 
       test('wrong type returns empty', () {
-        expect(pick(Object()).asListOrNull((it) => it.asInt()), isNull);
+        expect(pick(Object()).asListOrNull((it) => it.asIntOrThrow()), isNull);
       });
 
       test('map empty list to empty list', () {
@@ -278,7 +281,7 @@ void main() {
         expect(
           () => pick(data).asListOrNull((pick) => Person.fromPick(pick)),
           throwsA(pickException(containing: [
-            'required value at location list index 1 in pick(json, 1 (absent), "name") is absent.'
+            'Expected a non-null value but location list index 1 in pick(json, 1 (absent), "name") is absent.'
           ])),
         );
       });